requirements.txt
coloredlogs>=5.2
h5py>=2.6.0
IPython>=5.1.0
keras>=2.0.1
librosa>=0.5.0
matplotlib>=2.0.0
msgpack-python>=0.4.8
numpy>=1.9.2
pydot-ng >= 1.0.0
pyyaml>=3.11
scikit-learn>=0.16.1
scipy>=0.19.0
sed_eval>=0.1.2
six>=1.10.0
soundfile>=0.9.0
theano>=0.9.0rc4
tqdm >=4.11.2